configure-technirouter: support part 2 and launch a specific script

+++ Project2/configure-technirouter/configure-router.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+DIR=`pwd`
 dest="192.168.1.1"
 if [ $# -gt 0 ]; then
 	dest=$1;
@@ -27,11 +28,21 @@ sshcmd "opkg install *.ipk"
 
 
 echo
+read -p "Which part? (1, 2) " PART
+PARTDIR=../part$PART
 echo "Here is the list of routers"
-ls ../part1
+ls $PARTDIR
 read -p "Which one do you want? " RouterDir
-test -d ../part1/$RouterDir && echo "Copy in progress" || echo "Wrong dir"
-scpcmd -r ../part1/$RouterDir/* root@$dest:/
+test -d $PARTDIR/$RouterDir && echo "Copy in progress" || (echo "Wrong dir" && exit 1)
+
+# configure script
+if test -f "$PARTDIR/configure.sh"; then
+	cd $PARTDIR
+	./configure.sh $RouterDir tmp
+	RouterDir=tmp
+	cd $DIR
+fi
+scpcmd -r $PARTDIR/$RouterDir/* root@$dest:/
 
 echo "Install done - Restart the router"
 sshcmd "reboot"
